@@ -1,58 +1,33 @@
-import Register from './components/Register';
-import Login from './components/Login';
-import Home from './components/Home';
-import Layout from './components/Layout';
-import Editor from './components/Editor';
-import Admin from './components/Admin';
-import Missing from './components/Missing';
-import Unauthorized from './components/Unauthorized';
-import Lounge from './components/Lounge';
-import LinkPage from './components/LinkPage';
-import RequireAuth from './components/RequireAuth';
-import PersistLogin from './components/PersistLogin';
-import { Routes, Route } from 'react-router-dom';
-
-const ROLES = {
-  'User': 1,
-  'Editor': 1984,
-  'Admin': 42
-}
+import 'bootstrap/dist/css/bootstrap.min.css';
+import './App.css';
+import React from 'react';
+import Login from './comps/Login';
+import Pass from './comps/Pass';
+import Register from './comps/Register';
+import Navigation from './comps/Navigation';
+import Admin from './comps/Admin';
+import $ from 'jquery'
+import {
+  Routes,
+  Route,
+  HashRouter,
+} from "react-router-dom";
 
 function App() {
-
   return (
-    <Routes>
-      <Route path="/" element={<Layout />}>
-        {/* public routes */}
-        <Route path="login" element={<Login />} />
-        <Route path="register" element={<Register />} />
-        <Route path="linkpage" element={<LinkPage />} />
-        <Route path="unauthorized" element={<Unauthorized />} />
-
-        {/* we want to protect these routes */}
-        <Route element={<PersistLogin />}>
-          <Route element={<RequireAuth allowedRoles={[ROLES.User]} />}>
-            <Route path="/" element={<Home />} />
-          </Route>
-
-          <Route element={<RequireAuth allowedRoles={[ROLES.Editor]} />}>
-            <Route path="editor" element={<Editor />} />
-          </Route>
-
-
-          <Route element={<RequireAuth allowedRoles={[ROLES.Admin]} />}>
-            <Route path="admin" element={<Admin />} />
-          </Route>
-
-          <Route element={<RequireAuth allowedRoles={[ROLES.Editor, ROLES.Admin]} />}>
-            <Route path="lounge" element={<Lounge />} />
-          </Route>
-        </Route>
-
-        {/* catch all */}
-        <Route path="*" element={<Missing />} />
-      </Route>
-    </Routes>
+    <div className="App"><script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/crypto-js/3.1.2/rollups/aes.js"></script>
+    
+    <HashRouter basename="/">
+      <Routes>
+        <Route path='/' element={<Navigation></Navigation>} />
+        <Route path='/login' element={<Login/>} />
+        <Route path='/register' element={<Register/>} />
+        <Route path='/pass/:usu_id' element={<Pass  /> } />
+        <Route path='/admin/:token' element={<Admin  /> } />
+      </Routes>
+    </HashRouter>
+    <p id="logId" hidden></p>
+    </div>
   );
 }
 
